[csim] Added utility_sve.hpp
